feat: update theme colors and add logo font
This commit updates the primary and secondary theme colors across the application to use a new color palette, 'Driftwood' and 'Terracotta' respectively.

This change was made to improve the overall aesthetic of the application and align with updated branding guidelines.

Additionally, this commit introduces a new 'logo' font (Grandstander) and registers it with google fonts.