refactor: add ensurePublished state to StoryFactory
This commit introduces an `ensurePublished` state to the `StoryFactory`.

This allows for easily creating published stories in tests and seeders. Previously, the `published_at` attribute was randomly set to a date within the current month, and there was no easy way to ensure a story was published using the factory.

The new `ensurePublished` method sets the `status` to `Publish` and `published_at` to the current time using `Carbon::now()`.