test: add tests for edit action authorization in ViewStory component
This commit adds tests to verify that the edit action in the ViewStory Livewire component is correctly authorized based on the user's permissions and the story's creator.

The tests cover the following scenarios:

- A user who is the creator of the story should see the edit action and be authorized to use it.
- A user who is not the creator of the story should see the edit action, but should not be authorized to use it.
- A user with the "view_all_story" permission should see the edit action and be authorized to use it, even if they are not the creator of the story.

These tests ensure that the edit action is only available to users who have the necessary permissions to modify the story, enhancing the security and integrity of the application.