feat: indicate if user has replied to comment
This commit introduces a visual cue to indicate whether the logged-in user has already replied to a particular story comment.

This is achieved by:

- Adding a `hasUserReplied` boolean property to the `StoryCommentCard` Livewire component.
- In the `mount` method, checking if the user is logged in and if so, querying the comment's replies to see if any were created by the current user.
- Passing this information to the view.
- Updating the `filament::button` component for replies to show as filled if the user has replied, and outlined if not.

This improves the user experience by providing immediate feedback on their interaction with the comment thread, making it easier to track which comments they have already responded to.