feat: allow hiding actions in story comment card
This commit introduces the ability to hide the edit/delete actions in the StoryCommentCard component.

Previously, the actions were always displayed. This change adds a `$showActions` property to the component, which can be set to `false` to hide the actions. This is useful in contexts where the actions are not relevant, such as when displaying a parent comment in the StoryCommentResource form.