feat: add indexing to vote columns in story comments
This commit modifies the `story_comments` table migration to add indexes to the `upvote_count`, `downvote_count`, `vote_count`, and `vote_score` columns.

The addition of indexes is intended to improve the performance of queries that involve filtering or sorting by these vote-related columns. The `vote_score` column has also been updated to use `decimal(8, 2)` to allow for fractional scores, and an index has been added to this column as well.

The `down` method of the migration has been updated to correctly drop the added indexes before dropping the columns.