feat: create language_story table
This commit introduces a new database migration to create the `language_story` table.

This table serves as a pivot table for the many-to-many relationship between languages and stories. It includes foreign keys for `language_id` and `story_id`, both of which are defined as `foreignUlid` and set to cascade on delete. The table also includes standard `timestamps`.