feat: add Language model
This commit introduces the `Language` Eloquent model.

The model is located at `app/Models/Language.php` and extends `Illuminate\Database\Eloquent\Model`. It utilizes the `HasUlids` trait for ULID generation and has `code` and `name` as fillable properties. This model will be used to represent language entries within the application.