feat: automatically sync story languages
This commit introduces a new `StoryObserver` to automatically manage the many-to-many relationship between `Story` and `Language` models.

The observer's `saved` method is triggered whenever a story is saved. It checks if the `content` attribute has been modified. If it has, the observer extracts all language codes present in the translated `content` field.

It then synchronizes the `languages` relationship of the story with the identified language codes. This means new languages found in the content will be attached, and languages previously associated with the story but no longer present in the content will be detached.

Additionally, the observer logs a warning if any language codes found in the story's content do not exist in the `languages` table, helping to identify potential data inconsistencies. The `Story` model is now annotated with `#[ObservedBy([StoryObserver::class])]` to register this observer.