feat: add factory for Language model
This commit introduces a new factory for the `Language` Eloquent model. This factory will be used to generate dummy data for the `Language` model, which is essential for testing and development purposes.

The factory, located at `database/factories/LanguageFactory.php`, defines the default state for a `Language` instance, generating a unique language code and a placeholder country name for the language name.

The `HasFactory` trait has been added to the `Language` model to enable the use of this factory. The PHPDoc block has also been updated to include the `factory()` method signature.